+++ src/main/java/slimeknights/tconstruct/library/fluid/FluidTooltipHandler.java
@@ -26,9 +26,7 @@ import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraft.world.level.material.Fluid;
-import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import io.github.fabricators_of_create.porting_lib.util.FluidStack;
-import net.minecraftforge.fml.ModList;
 import slimeknights.mantle.recipe.ingredient.FluidIngredient;
 import slimeknights.mantle.util.JsonHelper;
 import slimeknights.mantle.util.RegistryHelper;
@@ -236,10 +234,10 @@ public class FluidTooltipHandler extends SimpleJsonResourceReloadListener {
    * @param tooltip    Tooltip to append information
    * @return  True if the amount is not in buckets
    */
-  public static boolean appendMaterialNoShift(Fluid fluid, int original, List<Component> tooltip) {
+  public static boolean appendMaterialNoShift(Fluid fluid, long original, List<Component> tooltip) {
     // if holding shift, skip specific units
     if (SafeClientAccess.getTooltipKey() != TooltipKey.SHIFT) {
-      int amount = original;
+      long amount = original;
       amount = INSTANCE.getUnitList(fluid).getText(tooltip, amount);
       MILLIBUCKET.getText(tooltip, amount);
       return INSTANCE.listCache.get(fluid) != INSTANCE.fallback;
@@ -279,7 +277,7 @@ public class FluidTooltipHandler extends SimpleJsonResourceReloadListener {
    * @param amount   Fluid amount
    * @param tooltip  Tooltip to append information
    */
-  public static void appendNamedList(ResourceLocation id, int amount, List<Component> tooltip) {
+  public static void appendNamedList(ResourceLocation id, long amount, List<Component> tooltip) {
     amount = INSTANCE.getUnitList(id).getText(tooltip, amount);
     appendBuckets(amount, tooltip);
   }
@@ -299,7 +297,7 @@ public class FluidTooltipHandler extends SimpleJsonResourceReloadListener {
   @RequiredArgsConstructor
   public static class FluidUnit {
     private final String key;
-    private final int needed;
+    private final long needed;
 
     /**
      * Gets the display text for this fluid entry
@@ -328,7 +326,7 @@ public class FluidTooltipHandler extends SimpleJsonResourceReloadListener {
     }
 
     /** Applies the text of all child units */
-    public int getText(List<Component> tooltip, int amount) {
+    public long getText(List<Component> tooltip, long amount) {
       if (units != null) {
         for (FluidUnit unit : units) {
           amount = unit.getText(tooltip, amount);
